requirements.txt
# common
numpy==2.2.4
scipy==1.15.2
# jax related
jax==0.5.3
optax==0.2.4
# visualization
matplotlib==3.10.1
PyQt5==5.15.11
PyYAML==6.0.2
